mctp-netlink: Store interface physical address in linkmap
Normally, extended addressing is used to respond to requests. We can
take the physical address len from the requester's sockaddr.
However, when we want to send extended address on an interface,
we do not have that information in linkmap.

This stores interface address from netlink inside linkmap.

Tested: Discovery Notify message is filled with correct address
(see next commit)

Signed-off-by: Khang Nguyen Duy <khangng@amperecomputing.com>